integrate mtp config version
