+++ src/constants.js
@@ -316,7 +316,7 @@ export const version = {
   backend: config.backendVersion,
   backendURL: 'https://github.com/CBIIT/ppdc-otp-backend/releases',
   changeLogPage: '/change-log',
-  fdaPmtlData: '3.0',
+  versionConfigURL: `${config.mtpConfig}/front-end/assets/data/version.json`,
 };
 
 export const mtpPageNames = {
